requirements.txt
openai
google-genai
anthropic
python-dotenv
tqdm